Specify library dependencies in library.properties
Specifying the library dependencies in the depends field of library.properties causes the Arduino Library Manager (Arduino IDE 1.8.10 and newer) to offer to install any missing dependencies during installation of this library. "arduino-cli lib install" will automatically install the dependencies (arduino-cli 0.7.0 and newer).

NOTE: I did not add the ArduinoJson dependency because the Library Manager dependencies system currently only allows installing the newest version of the dependencies and this library is only compatible with ArduinoJson 5.x.x. In the event that this library is fixed to work with the modern ArduinoJson API, or the Library Manager dependencies system is improved to support dependency version specifications, ArduinoJson should be added to the depends field.